model relationship between users and posts, add user's posts to their dashboard, access control to only allow the correct user to edit or delete a post

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,11 @@ use DB;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         $posts = Post::orderBy('created_at', 'desc')->paginate(10);
@@ -45,6 +50,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
   
@@ -67,8 +78,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
 
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        
+        $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
 }
